feat: sign out button is now on profile page
